[component] fix bucket aggregation
This modifies bucket aggregations so that they don't include the first bucket in the range since it
will never be provided data due to how it is queried.

Since the bucket selection algorithm needed to change because of this, it was simplified and test
cases added to clarify how it works.